+++ lib/PaygreenSdk/Payment/V3/Client.php
@@ -9,7 +9,7 @@ use Paygreen\Sdk\Payment\V3\Model\Instrument;
 use Paygreen\Sdk\Payment\V3\Model\PaymentOrder;
 use Paygreen\Sdk\Payment\V3\Request\Authentication\AuthenticationRequest;
 use Paygreen\Sdk\Payment\V3\Request\Buyer\BuyerRequest;
-use Paygreen\Sdk\Payment\V3\Request\Config\ConfigRequest;
+use Paygreen\Sdk\Payment\V3\Request\PaymentConfig\PaymentConfigRequest;
 use Paygreen\Sdk\Payment\V3\Request\Instrument\InstrumentRequest;
 use Paygreen\Sdk\Payment\V3\Request\PaymentOrder\OrderRequest;
 use Psr\Http\Message\ResponseInterface;
@@ -55,9 +55,9 @@ class Client extends \Paygreen\Sdk\Core\Client
      *
      *@return ResponseInterface
      */
-    public function getConfig()
+    public function listPaymentConfig()
     {
-        $request = (new ConfigRequest($this->requestFactory, $this->environment))->getGetRequest();
+        $request = (new PaymentConfigRequest($this->requestFactory, $this->environment))->getGetRequest();
         $this->setLastRequest($request);
 
         $response = $this->sendRequest($request);
